BACKPORT: sched/fair: Re-Introduce an energy estimation helper function
The previous implementation can be found in below commit.
https://github.com/TheHitMan7/kernel_xiaomi_msm8953/commit/dc9b764b5e5bbd0e68c7fbb6c3502b793270e765

It has been modified for the following reason:

BACKPORT: Reworked cpu_util_next() to work with
WALT. Removed the call to schedutil_energy_util() which means we only
look at CFS signals for the energy calculation. That last point
shouldn't make a huge difference as RT and DL pressure should hopefully
be small, and we don't have the PELT signals for these classes in 4.9.

But now, we have both PELT and UTIL_EST implementation in Linux kernel 3.18.y
So add back the original mainline implementation for estimated utilization.

Signed-off-by: TheHitMan7 <krtik.vrma@gmail.com>